+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## 2.55.0 (2026-04-08)
+
+### Features
+
+- **firehose**: Add Firehose client (#4182)
+
+### Bug Fixes
+
+- using Hub category to fire configured event (#3995)
+- **kinesis**: enable full flush for cache (#4169)
+
+## 2.54.1 (2026-03-11)
+
+### Bug Fixes
+
+- **api**: populate the auth mode when parsing the request response. So AppSyncListProvider.getNextPage() will have the auth mode. (#4156)
+
+## 2.54.0 (2026-03-04)
+
+### Features
+
+- **kinesis**: fix retry logic for request-level errors (#4167)
+- **kinesis**: add Kinesis Client (#4141)
+- **foundation**: Add Foundation libraries (#4150)
+
 ## 2.53.3 (2026-01-15)
 
 ### Bug Fixes